+++ packages/hundredandten-automation-engineadapter/pyproject.toml
@@ -3,9 +3,9 @@ requires = ["uv_build>=0.11.2,<0.12"]
 build-backend = "uv_build"
 
 [project]
-name = "hundredandten-automation-naive"
-version = "0.0.1.dev2"
-description = "Naive strategy player for the game Hundred and Ten"
+name = "hundredandten-automation-engineadapter"
+version = "0.0.1.dev0"
+description = "Engine adapter for wiring automation strategies to the Hundred and Ten game engine"
 readme = "README.md"
 requires-python = ">=3.12"
 license = {text = "MIT"}
@@ -16,14 +16,16 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
 ]
-dependencies = ["hundredandten-state>=0.0.0,<1.0.0", "hundredandten-deck>=0.0.0,<1.0.0"]
+dependencies = [
+    "hundredandten-state>=0.0.0,<1.0.0",
+    "hundredandten-engine>=0.0.0,<1.0.0",
+    "hundredandten-deck>=0.0.0,<1.0.0",
+]
 
 [dependency-groups]
 test = [
-    "hundredandten-automation-engineadapter>=0.0.0,<1.0.0",
-    "hundredandten-engine>=0.0.0,<1.0.0",
     "hundredandten-testing>=0.0.0,<1.0.0",
 ]
 
 [tool.uv.build-backend]
-module-name = "hundredandten.automation.naive"
+module-name = "hundredandten.automation.engineadapter"
