+++ packages/hundredandten-state/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "uv_build"
 
 [project]
 name = "hundredandten-state"
-version = "0.0.1.dev3"
+version = "0.0.1.dev4"
 description = "Player observation layer for the game Hundred and Ten"
 readme = "README.md"
 requires-python = ">=3.12"
@@ -16,7 +16,7 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
 ]
-dependencies = ["hundredandten-deck>=0.0.0,<1.0.0", "hundredandten-engine>=0.0.0,<1.0.0"]
+dependencies = ["hundredandten-deck>=0.0.0,<1.0.0"]
 
 [tool.uv.build-backend]
 module-name = "hundredandten.state"
