+++ pyproject.toml
@@ -28,14 +28,26 @@ lint.isort.combine-as-imports = true
 target-version = ["py314"]
 
 [tool.pytest.ini_options]
-testpaths = ["packages/hundredandten-engine", "packages/hundredandten-state", "packages/hundredandten-automation-naive", "packages/hundredandten-deck"]
+testpaths = [
+    "packages/hundredandten-engine",
+    "packages/hundredandten-state",
+    "packages/hundredandten-automation-naive",
+    "packages/hundredandten-automation-engineadapter",
+    "packages/hundredandten-deck"
+]
 addopts = [
     "--import-mode=importlib",
 ]
 
 [tool.coverage.run]
-source = ["hundredandten.engine", "hundredandten.state", "hundredandten.automation.naive", "hundredandten.deck"]
-omit = ["*/tests/*", "*/hundredandten/testing/*", "*/hundredandten-testing/*"]
+source = [
+    "hundredandten.engine",
+    "hundredandten.state",
+    "hundredandten.automation.naive",
+    "hundredandten.automation.engineadapter",
+    "hundredandten.deck"
+]
+omit = ["*/tests/*", "*/hundredandten-testing/*"]
 branch = true
 
 [tool.coverage.report]
@@ -48,6 +60,7 @@ check-url = "https://pypi.org/simple/"
 hundredandten-engine = { workspace = true }
 hundredandten-state = { workspace = true }
 hundredandten-automation-naive = { workspace = true }
+hundredandten-automation-engineadapter = { workspace = true }
 hundredandten-testing = { workspace = true }
 hundredandten-deck = { workspace = true }
